+++ BrainNetworksInPython/datasets/__init__.py
@@ -1,0 +1,4 @@
+"""
+Datasets module
+"""
+from . import NSPN_WhitakerVertes_PNAS2016
